+++ lib/fetcher.dart
@@ -5,7 +5,7 @@ export 'src/config/fetcher_config.dart';
 export 'src/exceptions/fetch_exception.dart';
 export 'src/models/fetch_error_data.dart';
 
-export 'src/async_form.dart';
+export 'src/submit_form_builder.dart';
 export 'src/submit_builder.dart';
 export 'src/async_edit_builder.dart';
 export 'src/fetch_builder.dart';